fix: boardLiveTime 초 단위로 변환하여 반환

+++ backend/src/main/java/com/trend_now/backend/board/dto/BoardInfoDto.java
@@ -12,7 +12,7 @@ public class BoardInfoDto {
     private Long boardId;
     private String boardName;
     private long boardLiveTime;
-    private long boardExpiredTime;
+    private Double boardExpiredTime;
     private double score;
     private String summary;
 }
